chore: added proper user verification for delete user endpoint

+++ src/controllers/auth.controller.js
@@ -206,20 +206,27 @@ async function handleChangePassword(req,res){
 
 async function handleDeleteUser(req, res) {
     try{
-        const {id} = req.body
+        const {userId: id} = req.body
 
         const user = await UserModel.countDocuments({_id: id});
-
+        if(!user){
+            return res.status(404).json({
+                message: "user not found",
+                success: false,
+                statusCode: 404
+            });
+        }
         const deletedUser = await UserModel.deleteOne({_id: id });
 
     return res.status(200).json({
         message: "user deleted successfully",
         deletedUser,
+        success: true,
         statusCode: 200
     });
     
 }    catch (error) {
-    // console.log(error)
+    console.log(error)
     res.status(400).json({
     success: false,
     message: "something went wrong",
